refactor: split multi-type files to comply with SA1402
Each public type now lives in its own file, matching the file name (SA1649).
Previously suppressed SA1402/SA1649 rules are now active in all src/ projects.

Splits:
- NexJob: Models.cs → JobId, JobPriority, JobStatus, JobRecord,
          RecurringConcurrencyPolicy, RecurringJobRecord
- NexJob: DashboardModels.cs → JobMetrics, HourlyThroughput, PagedResult,
          JobFilter, QueueMetrics
- NexJob: Configuration.cs → NexJobOptions, NexJobServiceCollectionExtensions
- NexJob: Extensibility.cs → ThrottleAttribute, IJobMigration
- NexJob.Postgres: RowMappings.cs → JobRow, RecurringJobRow
- NexJob.MongoDB: Documents.cs → JobDocument, RecurringJobDocument
- NexJob.MongoDB: JobIdSerializer.cs → JobIdSerializer, NullableJobIdSerializer
- NexJob.Dashboard: DashboardMiddleware.cs → DashboardApplicationBuilderExtensions

SA1402/SA1649 kept suppressed only for tests/ (fixtures alongside test classes)
and samples/ (input records co-located with job classes for demo readability).
