+++ src/NexJob/PagedResult.cs
@@ -1,0 +1,21 @@
+namespace NexJob;
+
+/// <summary>A page of results from a paginated query.</summary>
+/// <typeparam name="T">The type of items returned in each page.</typeparam>
+public sealed class PagedResult<T>
+{
+    /// <summary>Items on the current page.</summary>
+    public IReadOnlyList<T> Items { get; init; } = [];
+
+    /// <summary>Total number of items across all pages.</summary>
+    public int TotalCount { get; init; }
+
+    /// <summary>1-based current page number.</summary>
+    public int Page { get; init; }
+
+    /// <summary>Maximum number of items per page.</summary>
+    public int PageSize { get; init; }
+
+    /// <summary>Total number of pages.</summary>
+    public int TotalPages => PageSize > 0 ? (int)Math.Ceiling((double)TotalCount / PageSize) : 0;
+}
